+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/data/GetMessagesAfterDto.java
@@ -19,4 +19,5 @@ public class GetMessagesAfterDto {
     List<HtmlMessage> htmlMsgList;
     List<MessageMetaDataDto> metaMsgList;
     Integer pagesAvailable;
+    Integer totalMsgCount;
 }